rename comment_entry.rs -> comments/mod.rs
